inline-help: move selectors into proper folder.
