+++ client/blocks/inline-help/index.jsx
@@ -23,7 +23,7 @@ import isHappychatOpen from 'state/happychat/selectors/is-happychat-open';
 import hasActiveHappychatSession from 'state/happychat/selectors/has-active-happychat-session';
 import AsyncLoad from 'components/async-load';
 import { showInlineHelpPopover, hideInlineHelpPopover } from 'state/inline-help/actions';
-import isInlineHelpPopoverVisible from 'state/selectors/is-inline-help-popover-visible';
+import isInlineHelpPopoverVisible from 'state/inline-help/selectors/is-inline-help-popover-visible';
 import isInlineHelpVisible from 'state/selectors/is-inline-help-visible';
 
 /**
