+++ client/blocks/inline-help/inline-help-rich-result.jsx
@@ -24,7 +24,7 @@ import {
 import { Button } from '@automattic/components';
 import { decodeEntities, preventWidows } from 'lib/formatting';
 import { recordTracksEvent } from 'state/analytics/actions';
-import getSearchQuery from 'state/selectors/get-search-query';
+import getSearchQuery from 'state/inline-help/selectors/get-search-query';
 import { requestGuidedTour } from 'state/ui/guided-tours/actions';
 import { openSupportArticleDialog } from 'state/inline-support-article/actions';
