+++ client/blocks/inline-help/inline-help-search-card.jsx
@@ -13,10 +13,10 @@ import debugFactory from 'debug';
  */
 import { recordTracksEvent } from 'state/analytics/actions';
 import SearchCard from 'components/search-card';
-import getInlineHelpCurrentlySelectedLink from 'state/selectors/get-inline-help-currently-selected-link';
-import getSelectedResultIndex from 'state/selectors/get-selected-result-index';
-import isRequestingInlineHelpSearchResultsForQuery from 'state/selectors/is-requesting-inline-help-search-results-for-query';
-import getInlineHelpCurrentlySelectedResult from 'state/selectors/get-inline-help-currently-selected-result';
+import getInlineHelpCurrentlySelectedLink from 'state/inline-help/selectors/get-inline-help-currently-selected-link';
+import getSelectedResultIndex from 'state/inline-help/selectors/get-selected-result-index';
+import isRequestingInlineHelpSearchResultsForQuery from 'state/inline-help/selectors/is-requesting-inline-help-search-results-for-query';
+import getInlineHelpCurrentlySelectedResult from 'state/inline-help/selectors/get-inline-help-currently-selected-result';
 import {
 	requestInlineHelpSearchResults,
 	selectNextResult,
