+++ client/blocks/inline-help/popover.jsx
@@ -19,8 +19,8 @@ import Popover from 'components/popover';
 import InlineHelpSearchResults from './inline-help-search-results';
 import InlineHelpSearchCard from './inline-help-search-card';
 import InlineHelpRichResult from './inline-help-rich-result';
-import getSearchQuery from 'state/selectors/get-search-query';
-import getInlineHelpCurrentlySelectedResult from 'state/selectors/get-inline-help-currently-selected-result';
+import getSearchQuery from 'state/inline-help/selectors/get-search-query';
+import getInlineHelpCurrentlySelectedResult from 'state/inline-help/selectors/get-inline-help-currently-selected-result';
 import { getHelpSelectedSite } from 'state/help/selectors';
 import QuerySupportTypes from 'blocks/inline-help/inline-help-query-support-types';
 import InlineHelpContactView from 'blocks/inline-help/inline-help-contact-view';
