+++ client/components/data/query-inline-help-search/index.jsx
@@ -8,7 +8,7 @@ import { connect } from 'react-redux';
  * Internal dependencies
  */
 import { requestInlineHelpSearchResults } from 'state/inline-help/actions';
-import isRequestingInlineHelpSearchResultsForQuery from 'state/selectors/is-requesting-inline-help-search-results-for-query';
+import isRequestingInlineHelpSearchResultsForQuery from 'state/inline-help/selectors/is-requesting-inline-help-search-results-for-query';
 
 class QueryInlineHelpSearch extends Component {
 	UNSAFE_componentWillMount() {
