+++ client/my-sites/customer-home/cards/features/help-search/search-results.jsx
@@ -15,9 +15,9 @@ import { recordTracksEvent } from 'state/analytics/actions';
 import QueryInlineHelpSearch from 'components/data/query-inline-help-search';
 import PlaceholderLines from './placeholder-lines';
 import { decodeEntities, preventWidows } from 'lib/formatting';
-import getInlineHelpSearchResultsForQuery from 'state/selectors/get-inline-help-search-results-for-query';
-import getSelectedResultIndex from 'state/selectors/get-selected-result-index';
-import isRequestingInlineHelpSearchResultsForQuery from 'state/selectors/is-requesting-inline-help-search-results-for-query';
+import getInlineHelpSearchResultsForQuery from 'state/inline-help/selectors/get-inline-help-search-results-for-query';
+import getSelectedResultIndex from 'state/inline-help/selectors/get-selected-result-index';
+import isRequestingInlineHelpSearchResultsForQuery from 'state/inline-help/selectors/is-requesting-inline-help-search-results-for-query';
 import hasInlineHelpAPIResults from 'state/selectors/has-inline-help-api-results';
 
 import { setSearchResults, selectResult } from 'state/inline-help/actions';
