+++ client/state/inline-help/selectors/get-search-query.js
