+++ client/state/inline-help/selectors/get-selected-result-index.js
