+++ client/state/selectors/test/is-inline-help-popover-visible.js
@@ -6,7 +6,7 @@ import { expect } from 'chai';
 /**
  * Internal dependencies
  */
-import isInlineHelpPopoverVisible from '../is-inline-help-popover-visible';
+import isInlineHelpPopoverVisible from './is-inline-help-popover-visible';
 
 describe( '#isInlineHelpPopoverVisible()', () => {
 	test( 'should return if the popover is visible', () => {
